Harden block template parsing edge cases
Malformed varints were previously decoded through read_varint without propagating its failure status, so truncated, overflowing, or non-canonical values could leave low parsed values and pass later validation. The binary archive reader now decodes varints directly and sets failbit on EOF, overflow, and non-canonical encodings so existing serialization macros fail closed.

The local Keccak wrapper also accepted size_t lengths and narrowed them to int before hashing, which could turn very large blob lengths into negative memcpy sizes. The Keccak API now carries size_t lengths through the implementation.

Finally, address_decode and address_decode_integrated decoded uint64 prefixes but returned only the low uint32 bits to JavaScript. Prefixes that cannot be represented by the exported 32-bit API are now rejected instead of being silently truncated.

Verified with: npm rebuild; npm test.

+++ src/crypto/keccak.h
@@ -4,6 +4,7 @@
 #ifndef KECCAK_H
 #define KECCAK_H
 
+#include <stddef.h>
 #include <stdint.h>
 #include <string.h>
 
@@ -16,11 +17,11 @@
 #endif
 
 // compute a keccak hash (md) of given byte length from "in"
-int keccak(const uint8_t *in, int inlen, uint8_t *md, int mdlen);
+int keccak(const uint8_t *in, size_t inlen, uint8_t *md, int mdlen);
 
 // update the state
 void keccakf(uint64_t st[25], int norounds);
 
-void keccak1600(const uint8_t *in, int inlen, uint8_t *md);
+void keccak1600(const uint8_t *in, size_t inlen, uint8_t *md);
 
 #endif
